Safari-611.1.10/LayoutTests/js/symbol-in-map-expected.txt

This tests Symbols work in ES6 Map.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS map.size is 1
PASS map.get(symbol) is "Cappuccino"
PASS map.has(Symbol('Cocoa')) is false
PASS map.get(Symbol('Cocoa')) is undefined
PASS map.has('Cocoa') is false
PASS map.get('Cocoa') is undefined
PASS map.has(symbol2) is false
PASS map.get(symbol2) is undefined
PASS map.size is 2
PASS map.get(symbol) is "Cappuccino"
PASS map.get(symbol2) is "Kilimanjaro"
PASS map.has(Symbol('Matcha')) is false
PASS map.get(Symbol('Matcha')) is undefined
PASS map.has('Matcha') is false
PASS map.get('Matcha') is undefined
PASS map.has(symbol2) is false
PASS map.get(symbol2) is undefined
PASS map.has(symbol) is true
PASS map.get(symbol) is "Cappuccino"
PASS map.size is 1
PASS key is symbol
PASS value is "Cappuccino"
PASS successfullyParsed is true

TEST COMPLETE
